requirements.txt
# Core Dependencies
flask==2.3.3
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1

# Browser Automation & Anti-Detection
undetected-chromedriver==3.5.4
playwright==1.40.0
selenium==4.15.2

# AI & NLP
groq==0.4.1
openai==1.3.7
langchain==0.0.335
spacy==3.7.2
transformers==4.35.2
torch==2.1.1

# Document Processing
PyMuPDF==1.23.8
python-docx==1.1.0
pypdf2==3.0.1
pdfplumber==0.9.0

# Data Processing
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Database
sqlite3
sqlalchemy==2.0.23
tinydb==4.8.0

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
apscheduler==3.10.4
fake-useragent==1.4.0
python-dateutil==2.8.2
pydantic==2.5.0

# Email & Notifications
smtplib
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
